requirements.txt
instaloader==4.9.6
python-telegram-bot==13.7
requests==2.27.1